move types into apply.go
